requirements.txt
google-genai
redisvl